requirements.txt
numpy<2.0
flask==3.0.0
transformers==4.36.2
torch==2.1.2
waitress==2.1.2
python-dotenv==1.0.0
pytest==7.4.3
